+++ example-app/src/app/page.tsx
@@ -6,7 +6,7 @@ import Script from 'next/script'
 export default function Home() {
   return (
     <>
-      <Script src="https://js.credova.com/v1.1.1" />
+      <Script src="https://js.publicsquare.com/" />
       <Script src="https://js.basistheory.com/" />
       <main className="flex min-h-screen flex-col items-center justify-between p-24">
         <div className="relative z-[-1] flex place-items-center before:absolute before:h-[300px] before:w-full before:-translate-x-1/2 before:rounded-full before:bg-gradient-radial before:from-white before:to-transparent before:blur-2xl before:content-[''] after:absolute after:-z-20 after:h-[180px] after:w-full after:translate-x-1/3 after:bg-gradient-conic after:from-sky-200 after:via-blue-200 after:blur-2xl after:content-[''] before:dark:bg-gradient-to-br before:dark:from-transparent before:dark:to-blue-700 before:dark:opacity-10 after:dark:from-sky-900 after:dark:via-[#0141ff] after:dark:opacity-40 sm:before:w-[480px] sm:after:w-[240px] before:lg:h-[360px]">
@@ -27,7 +27,7 @@ export default function Home() {
 
         <div className="mb-32 flex justify-center text-center lg:mb-0 lg:w-full lg:max-w-5xl lg:text-left">
           <a
-            href="https://docs.credova.com"
+            href="https://developers.publicsquare.com"
             className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
             target="_blank"
             rel="noopener noreferrer"
@@ -52,7 +52,7 @@ export default function Home() {
               </span>
             </h2>
             <p className="m-0 max-w-[30ch] text-sm opacity-50">
-              Find in-depth information about Credova features and SDK.
+              Find in-depth information about PublicSquare features and SDK.
             </p>
           </a>
         </div>