+++ js-sdk/src/cards/Cards.ts
@@ -1,35 +1,35 @@
 import { CardCreateResponse, CardsCreateInput } from '@/types/sdk/cards'
-import { Credova } from '..'
-import { ELEMENTS_CREDOVA_CARDS_NO_POINTER_MESSAGE } from '@/constants'
+import { PublicSquare } from '..'
+import { ELEMENTS_CARDS_NO_POINTER_MESSAGE } from '@/constants'
 import { transformCreateCardInput } from '@/utils'
 import { validateCreateCardInput } from '@/validators'
 
-export class CredovaCards {
-  private _credova: Credova
+export class PublicSquareCards {
+  private _publicSquare: PublicSquare
 
-  constructor(credovaPointer: Credova) {
-    if (!credovaPointer) {
-      throw Error(ELEMENTS_CREDOVA_CARDS_NO_POINTER_MESSAGE)
+  constructor(publicSquarePointer: PublicSquare) {
+    if (!publicSquarePointer) {
+      throw Error(ELEMENTS_CARDS_NO_POINTER_MESSAGE)
     }
-    this._credova = credovaPointer
+    this._publicSquare = publicSquarePointer
   }
 
   public create(input: CardsCreateInput): Promise<CardCreateResponse> {
-    if (!this._credova._apiKey) {
+    if (!this._publicSquare._apiKey) {
       throw new Error('apiKey must be sent at initialization')
-    } else if (!this._credova.bt || !this._credova.bt.client) {
-      throw new Error('Credova JS has not be initialized yet')
+    } else if (!this._publicSquare.bt || !this._publicSquare.bt.client) {
+      throw new Error('PublicSquare JS has not be initialized yet')
     } else {
       const validatedInput = validateCreateCardInput(input)
-      return this._credova.bt.client
+      return this._publicSquare.bt.client
         .post(
           'https://api.basistheory.com/proxy',
           transformCreateCardInput(validatedInput),
           {
             headers: {
               'Content-Type': 'application/json',
-              'X-API-KEY': this._credova._apiKey,
-              'BT-PROXY-KEY': this._credova._proxyKey,
+              'X-API-KEY': this._publicSquare._apiKey,
+              'BT-PROXY-KEY': this._publicSquare._proxyKey,
             }
           }
         )