+++ js-sdk/src/constants.ts
@@ -1,20 +1,20 @@
 const ELEMENTS_INIT_ERROR_MESSAGE =
-  'Credova Elements was not properly initialized.'
+  'PublicSquare Elements was not properly initialized.'
 
 const ELEMENTS_TYPE_NOT_SUPPORTED =
-  'Credova Elements does not support the type of element specified.'
+  'PublicSquare Elements does not support the type of element specified.'
 
 const ELEMENTS_NOM_DOM_ERROR_MESSAGE =
-  'PublicSquare requires access to the DOM (Client-side only support).'
+  'Tried to load PublicSquare Elements in a non-DOM environment. Only client-side environments are supported.'
 
 const ELEMENTS_SCRIPT_LOAD_ERROR_MESSAGE =
-  'Credova Elements did not load properly.'
+  'PublicSquare Elements did not load properly.'
 
 const ELEMENTS_SCRIPT_UNKNOWN_ERROR_MESSAGE =
-  'There was an unknown error when loading Credova Elements. Check the console for details.'
+  'There was an unknown error when loading PublicSquare Elements. Check the console for details.'
 
-const ELEMENTS_CREDOVA_CARDS_NO_POINTER_MESSAGE =
-  'CredovaCards requires credovaPointer when initializing'
+const ELEMENTS_CARDS_NO_POINTER_MESSAGE =
+  'PublicSquareCards requires publicSquarePointer when initializing'
 
 const ELEMENTS_PUBLICSQUARE_ACH_NO_POINTER_MESSAGE =
   'PublicSquareACH requires psqPointer when initializing'
@@ -40,7 +40,7 @@ const CARD_ICON_POSITIONS = ['left', 'right', 'none'] as const
 const AUTOCOMPLETE_VALUES = ['off', 'on'] as const
 
 const API_ENDPOINTS = {
-  BASE_URL: 'https://collect-staging.credova.com',
+  BASE_URL: 'https://collect-staging.publicsquare.com',
   /**
    * Endpoint to save new cards to
    */
@@ -55,7 +55,7 @@ export {
   ELEMENTS_NOM_DOM_ERROR_MESSAGE,
   ELEMENTS_SCRIPT_LOAD_ERROR_MESSAGE,
   ELEMENTS_SCRIPT_UNKNOWN_ERROR_MESSAGE,
-  ELEMENTS_CREDOVA_CARDS_NO_POINTER_MESSAGE,
+  ELEMENTS_CARDS_NO_POINTER_MESSAGE,
   ELEMENTS_PUBLICSQUARE_ACH_NO_POINTER_MESSAGE,
   CARD_BRANDS,
   CARD_ICON_POSITIONS,