+++ js-sdk/src/types/sdk/index.ts
@@ -1,2 +1,2 @@
-export * from './credova'
-export * from './cards'
+export * from './publicsquare';
+export * from './cards';