+++ react-sdk/src/elements/CardNumberElement.tsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from 'react'
-import CredovaElement from './element'
-import Types from '@credova/elements-js/dist/types/sdk'
+import PublicSquareElement from './element'
+import Types from '@publicsquare/elements-js/types/sdk'
 
 type CardElementProps = {
   id: string
@@ -10,5 +10,5 @@ export const CardNumberElement = forwardRef<
   Types.CardNumberElement,
   CardElementProps
 >(function Component(props, ref) {
-  return <CredovaElement type="cardNumber" {...props} ref={ref} />
+  return <PublicSquareElement type="cardNumber" {...props} ref={ref} />
 })