+++ CHANGELOG.md
@@ -1,13 +1,41 @@
-# [1.1.0](https://github.com/credova-financial/credova-elements/compare/v1.0.0...v1.1.0) (2024-06-30)
+## [1.2.1](https://github.com/publicsquare-financial/publicsquare-elements/compare/v1.2.0...v1.2.1) (2024-09-14)
+
+
+### Bug Fixes
+
+* bump dependencies and js elements reference ([2ace4f7](https://github.com/publicsquare-financial/publicsquare-elements/commit/2ace4f7e3478453fc321d4c31402a61b992c2bd1))
+
+# [1.2.0](https://github.com/publicsquare-financial/publicsquare-elements/compare/v1.1.2...v1.2.0) (2024-09-14)
 
 
 ### Features
 
-* update readme ([be55540](https://github.com/credova-financial/credova-elements/commit/be55540a556c7417114487918b25315e15679268))
+* rename elements ([a946cea](https://github.com/publicsquare-financial/publicsquare-elements/commit/a946cea7b38c7e6a7ffd9a4c1991882f08bdbf0f))
+
+## [1.1.2](https://github.com/publicsquare-financial/publicsquare-elements/compare/v1.1.1...v1.1.2) (2024-08-08)
+
+
+### Bug Fixes
+
+* type export error, bump versions ([#26](https://github.com/publicsquare-financial/publicsquare-elements/issues/26)) ([c730497](https://github.com/publicsquare-financial/publicsquare-elements/commit/c730497540f2c98da556984ae9f1b78a3449dc86))
+
+## [1.1.1](https://github.com/publicsquare-financial/publicsquare-elements/compare/v1.1.0...v1.1.1) (2024-08-08)
+
+
+### Bug Fixes
+
+* js-sdk for <script /> tags ([#25](https://github.com/publicsquare-financial/publicsquare-elements/issues/25)) ([6fca433](https://github.com/publicsquare-financial/publicsquare-elements/commit/6fca433a32b5ce81825481f6cb30c1bc1a6961c5))
+
+# [1.1.0](https://github.com/publicsquare-financial/publicsquare-elements/compare/v1.0.0...v1.1.0) (2024-06-30)
+
+
+### Features
+
+* update readme ([be55540](https://github.com/publicsquare-financial/publicsquare-elements/commit/be55540a556c7417114487918b25315e15679268))
 
 # 1.0.0 (2024-06-28)
 
 
 ### Features
 
-* initial release ([a0ec752](https://github.com/credova-financial/credova-elements/commit/a0ec752f52c90cfd59afb0d1e64a2d5c75247845))
+* initial release ([a0ec752](https://github.com/publicsquare-financial/publicsquare-elements/commit/a0ec752f52c90cfd59afb0d1e64a2d5c75247845))